requirements.txt
flask
pymongo
flask-pymongo
python-dotenv